+++ packages/axios-extension/src/abap/catalog/v2-catalog-service.ts
@@ -111,7 +111,8 @@ export class V2CatalogService extends CatalogService {
             $format: 'json',
             $filter: `Title eq '${title}' and TechnicalServiceVersion eq ${version}`
         };
-        const response = await this.get<ODataServiceV2Info[]>(`/${this.entitySet}`, { params });
+        const requestPath = this.entitySet ? `/${this.entitySet}` : '/ServiceCollection';
+        const response = await this.get<ODataServiceV2Info[]>(requestPath, { params });
         const services = response.odata();
 
         if (services.length > 1) {
@@ -139,9 +140,6 @@ export class V2CatalogService extends CatalogService {
      * @returns service annotations
      */
     protected async getServiceAnnotations({ id, title, path }: FilterOptions): Promise<ODataServiceV2Info[]> {
-        if (!this.entitySet) {
-            await this.determineEntitySet();
-        }
         if (!id) {
             const ServiceConfig = await this.findService({ title, path });
             if (ServiceConfig) {
@@ -150,7 +148,7 @@ export class V2CatalogService extends CatalogService {
         }
         if (id) {
             const response = await this.get<ODataServiceV2Info[]>(
-                `/${this.entitySet}('${encodeURIComponent(id)}')/Annotations`,
+                `/ServiceCollection('${encodeURIComponent(id)}')/Annotations`,
                 {
                     params: { $format: 'json' }
                 }