+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.7.35
+
+### Patch Changes
+
+-   Updated dependencies [869c1c0d]
+    -   @sap-ux/axios-extension@1.13.1
+
+## 0.7.34
+
+### Patch Changes
+
+-   Updated dependencies [b2ee99fc]
+    -   @sap-ux/axios-extension@1.13.0
+
 ## 0.7.33
 
 ### Patch Changes