+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @sap-ux/fiori-elements-writer
 
+## 0.21.0
+
+### Minor Changes
+
+-   9188fe8b: fpm v4 removed sap.fe.templates in manifest.json, now has dependency on sap.fe.core
+    sap.ushell removed from ui5-application-writer, now loaded in fiori-elements and fiori-freestyle writers respectively
+
+### Patch Changes
+
+-   Updated dependencies [9188fe8b]
+    -   @sap-ux/ui5-application-writer@0.26.15
+    -   @sap-ux/fe-fpm-writer@0.24.15
+    -   @sap-ux/ui5-config@0.22.7
+    -   @sap-ux/odata-service-writer@0.20.2
+    -   @sap-ux/ui5-test-writer@0.3.3
+
+## 0.20.1
+
+### Patch Changes
+
+-   @sap-ux/ui5-config@0.22.6
+-   @sap-ux/odata-service-writer@0.20.1
+-   @sap-ux/ui5-application-writer@0.26.14
+-   @sap-ux/fe-fpm-writer@0.24.14
+-   @sap-ux/ui5-test-writer@0.3.3
+
+## 0.20.0
+
+### Minor Changes
+
+-   e9a007af: Add support for CAP project services and annotations, based on the service type.
+
+### Patch Changes
+
+-   Updated dependencies [e9a007af]
+    -   @sap-ux/odata-service-writer@0.20.0
+
+## 0.19.11
+
+### Patch Changes
+
+-   Updated dependencies [d2063073]
+    -   @sap-ux/fe-fpm-writer@0.24.14
+
+## 0.19.10
+
+### Patch Changes
+
+-   Updated dependencies [f69d72c1]
+    -   @sap-ux/fe-fpm-writer@0.24.13
+
+## 0.19.9
+
+### Patch Changes
+
+-   Updated dependencies [855a4ae2]
+    -   @sap-ux/fe-fpm-writer@0.24.12
+
 ## 0.19.8
 
 ### Patch Changes