+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.2.30
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.21.2
+
+## 0.2.29
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.21.1
+
+## 0.2.28
+
+### Patch Changes
+
+-   Updated dependencies [69b8d6de]
+    -   @sap-ux/project-access@1.21.0
+
+## 0.2.27
+
+### Patch Changes
+
+-   Updated dependencies [a7d78229]
+    -   @sap-ux/project-access@1.20.4
+
+## 0.2.26
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.20.3
+
 ## 0.2.25
 
 ### Patch Changes